+++ src/OpenTelemetry.OpAmp.Client/CHANGELOG.md
@@ -9,3 +9,7 @@
 * Added support for OpAMP WebSocket transport.
   ([#3064](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3064),
   [#3092](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3092))
+* Added support for heartbeat messages.
+  ([#3095](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3095))
+* Added EventSource `OpenTelemetry-OpAmp-Client` for diagnostics.
+  ([#3102](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/3102))